+++ MathComp2.h
@@ -12,35 +12,35 @@ class Matrix {
 
 private:
 
-	std::vector<std::vector<int>> data;
+	std::vector<std::vector<double>> data;
 	int rows;
 	int cols;
 
 public:
 
-	Matrix(const std::vector<std::vector<int>>& matrix_data);
-	Matrix(int m = 0, int n = 0);
+	Matrix(const std::vector<std::vector<double>>& matrix_data);
+	Matrix(double m = 0, double n = 0);
 
 	int getRows() const;
 	int getCols() const;
 
-	int& operator()(int i, int j);
+	double& operator()(double i, double j);
 
-	const int& operator()(int i, int j) const;
+	const double& operator()(double i, double j) const;
 
 	void readFromFile(const std::string& filename);
 	void readFromConsole();
 	void print(const std::string& name = "") const;
-	void printVector(const std::vector<int>& vec, const std::string& name = "");
+	void printVector(const std::vector<double>& vec, const std::string& name = "");
 
 	Matrix multiply(const Matrix& other) const;
 	Matrix operator*(const Matrix& other) const;
 
 	std::pair<Matrix, Matrix> luFactorization() const;
-	std::vector<int> solve(const std::vector<int>& b) const;
+	std::vector<double> solve(const std::vector<double>& b) const;
 
 	Matrix inverse() const;
 	Matrix transpose() const;
 	
-	std::vector<int> solveUsingInverse(const std::vector<int>& b) const;
+	std::vector<double> solveUsingInverse(const std::vector<double>& b) const;
 };
